pyproject.toml
[tool.poetry]
name = "blackjack"
version = "0.1.0"
description = ""
authors = ["Arnaldo Vitor <arnaldovitorbarros@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
attrs = "^23.1.0"
python-decouple = "^3.8"
numpy = "~1.24.4"
tensorflow-io-gcs-filesystem = "0.27.0"
tensorflow = "2.10.0"
loguru = "^0.7.2"


[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-cov = "*"
mypy = "*"
pip-audit = "*"
ruff = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
select = [
    "A",
    "ARG",
    "B",
    "C",
    "C4",
    "E",
    "ERA",
    "F",
    "I",
    "N",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "RET",
    "S",
    "T10",
    "UP",
    "W",
]
ignore = ["A003"]
target-version = "py310"

[tool.ruff.format]
quote-style = "single"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**" = ["ARG", "S"]
